# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ecommerce API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CATEGORIES ─────────────────────────────────────────────────────────────
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── COUPONS ────────────────────────────────────────────────────────────────
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Invalid coupon
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin only)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, stock, categoryId, images, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        stock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        variants:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isTrending:
          type: boolean
        isNewArrival:
          type: boolean
        isBestSeller:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        deliveryCharge:
          type: ["number", "null"]
        isDeliveryChargeApplicable:
          type: ["boolean", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, stock, images]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        discount:
          type: number
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        variants:
          type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isTrending:
          type: boolean
        isNewArrival:
          type: boolean
        isBestSeller:
          type: boolean
        deliveryCharge:
          type: number
        isDeliveryChargeApplicable:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        discount:
          type: number
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        variants:
          type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isTrending:
          type: boolean
        isNewArrival:
          type: boolean
        isBestSeller:
          type: boolean
        deliveryCharge:
          type: number
        isDeliveryChargeApplicable:
          type: boolean
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    WishlistItem:
      type: object
      required: [id, userId, productId, product]
      properties:
        id:
          type: integer
        userId:
          type: string
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer

    # ─── COUPONS ──────────────────────────────────────────────────────────────
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, isActive]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxDiscount:
          type: ["number", "null"]
        usageLimit:
          type: ["integer", "null"]
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxDiscount:
          type: number
        usageLimit:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
    CouponUpdate:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxDiscount:
          type: number
        usageLimit:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
    CouponValidateInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
        orderAmount:
          type: number

    # ─── ADMIN ────────────────────────────────────────────────────────────────
...