# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  /** @nullable */
  discount?: number | null;
  stock: number;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  images: string[];
  /** @nullable */
  variants?: string | null;
  tags?: string[];
  isFeatured?: boolean;
  isTrending?: boolean;
  isNewArrival?: boolean;
  isBestSeller?: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  /** @nullable */
  deliveryCharge?: number | null;
  /** @nullable */
  isDeliveryChargeApplicable?: boolean | null;
  createdAt: string;
}