# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug?: string;
  description?: string;
  price: number;
  comparePrice?: number;
  discount?: number;
  stock: number;
  categoryId?: number;
  images: string[];
  variants?: string;
  tags?: string[];
  isFeatured?: boolean;
  isTrending?: boolean;
  isNewArrival?: boolean;
  isBestSeller?: boolean;
  deliveryCharge?: number;
  isDeliveryChargeApplicable?: boolean;
}